requirements.txt
absl-py==1.1.0
appdirs==1.4.4
astunparse==1.6.3
audioread==2.1.9
cachetools==5.2.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
decorator==5.1.1
flatbuffers==1.12
gast==0.4.0
google-auth==2.9.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.47.0
h5py==3.7.0
idna==3.3
imageio==2.19.3
importlib-metadata==4.12.0
joblib==1.1.0
keras==2.7.0
Keras-Preprocessing==1.1.2
libclang==14.0.1
librosa==0.9.2
llvmlite==0.38.1
Markdown==3.3.7
networkx==2.6.3
numba==0.55.2
numpy==1.21.6
oauthlib==3.2.0
opt-einsum==3.3.0
packaging==21.3
Pillow==9.2.0
pooch==1.6.0
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio==0.2.11
pycparser==2.21
pyparsing==3.0.9
PyWavelets==1.3.0
requests==2.28.1
requests-oauthlib==1.3.1
resampy==0.3.1
rsa==4.8
scikit-image==0.19.3
scikit-learn==1.0.2
scipy==1.7.3
six==1.16.0
SoundFile==0.10.3.post1
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.7.0
tensorflow-estimator==2.7.0
tensorflow-io-gcs-filesystem==0.23.1
termcolor==1.1.0
threadpoolctl==3.1.0
tifffile==2021.11.2
typing_extensions==4.3.0
urllib3==1.26.10
Werkzeug==2.1.2
wrapt==1.14.1
zipp==3.8.0
